API/app/requirements.txt
flask
uwsgi
flask_cors
python-dotenv
selenium
google-cloud-translate